+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/enums/ModeType.java
@@ -2,5 +2,6 @@ package de.gematik.test.tiger.lib.enums;
 
 public enum ModeType {
   XML,
-  JSON
+  JSON,
+  JSON_SCHEMA
 }