+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/json/JsonSchemaChecker.java
@@ -1,0 +1,55 @@
+/*
+ * ${GEMATIK_COPYRIGHT_STATEMENT}
+ */
+
+package de.gematik.test.tiger.lib.json;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.networknt.schema.JsonSchema;
+import com.networknt.schema.JsonSchemaFactory;
+import com.networknt.schema.SpecVersion;
+import com.networknt.schema.ValidationMessage;
+import java.util.Set;
+
+/** Validates a given json string against a given json schema */
+public class JsonSchemaChecker {
+
+  private final ObjectMapper objectMapper = new ObjectMapper();
+
+  /**
+   * Checks that the json conforms to the given schema
+   *
+   * @param jsonToCheck json to check
+   * @param schema schema to check against
+   */
+  public void compareJsonToSchema(String jsonToCheck, String schema) {
+    try {
+      JsonNode jsonNode = objectMapper.readTree(jsonToCheck);
+      JsonNode schemaNode = objectMapper.readTree(schema);
+
+      JsonSchemaFactory factory = JsonSchemaFactory.getInstance(SpecVersion.VersionFlag.V202012);
+      JsonSchema jsonSchema = factory.getSchema(schemaNode);
+
+      Set<ValidationMessage> errors = jsonSchema.validate(jsonNode);
+
+      if (!errors.isEmpty()) {
+        throw new JsonSchemaAssertionError("Json does not conform to schema: " + errors);
+      }
+    } catch (JsonProcessingException e) {
+      throw new JsonSchemaAssertionError("Failed to process input json and/or input schema", e);
+    }
+  }
+
+  public static class JsonSchemaAssertionError extends AssertionError {
+
+    public JsonSchemaAssertionError(String message) {
+      super(message);
+    }
+
+    public JsonSchemaAssertionError(String message, Throwable cause) {
+      super(message, cause);
+    }
+  }
+}